README.md
# Android Demo

This is a simple game in android that demonstrates several design aspects in game development for android.  In this application users fly an drone across a field to collect points.  The game is multiplayer and after a player has reached 25 points they are designated the winner.  Players can choose to continue playing after the game is over.  

##What I learned from this project
*How to use Game Loop in android studio
*How to use Collision Events in Android Studio
*How to control animations of the applications user interface
*Using Images to make game animations

## Getting Started

To get started download a local copy of the game application.  

### Prerequisites

Android studio is required for running this application.  
Android api version 21 or higher is used during development.  
Java 8.



### Installing

Download zip files and and run the application.  

## Author

* **Chris Toplikar**  - (https://github.com/christoplikar)



## License

This project is licensed under the MIT License